# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Scalable multi-channel notification system",
        "title": "Notification Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/notifications": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notifications with filters and pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by channel",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by batch ID",
                        "name": "batch_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter from date (RFC3339)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to date (RFC3339)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20, max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create a notification",
                "parameters": [
                    {
                        "description": "Notification input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateNotificationInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Notification"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/batch": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create notifications in batch (up to 1000)",
                "parameters": [
                    {
                        "description": "Batch input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateBatchInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.BatchResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notification by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Notification"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Cancel a pending/queued notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "observability"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "observability"
                ],
                "summary": "Get real-time metrics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.BatchResult": {
            "type": "object",
            "properties": {
                "batch_id": {
                    "type": "string"
                },
                "queued": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "domain.Channel": {
            "type": "string",
            "enum": [
                "sms",
                "email",
                "push"
            ],
            "x-enum-varnames": [
                "ChannelSMS",
                "ChannelEmail",
                "ChannelPush"
            ]
        },
        "domain.CreateBatchInput": {
            "type": "object",
            "required": [
                "notifications"
            ],
            "properties": {
                "notifications": {
                    "type": "array",
                    "maxItems": 1000,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/domain.CreateNotificationInput"
                    }
                }
            }
        },
        "domain.CreateNotificationInput": {
            "type": "object",
            "required": [
                "channel",
                "recipient"
            ],
            "properties": {
                "channel": {
                    "enum": [
                        "sms",
                        "email",
                        "push"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Channel"
                        }
                    ]
                },
                "content": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "priority": {
                    "enum": [
                        "high",
                        "normal",
                        "low"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Priority"
                        }
                    ]
                },
                "recipient": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "domain.Notification": {
            "type": "object",
            "properties": {
                "batch_id": {
                    "type": "string"
                },
                "channel": {
                    "$ref": "#/definitions/domain.Channel"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/domain.Priority"
                },
                "provider_msg_id": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "retry_count": {
                    "type": "integer"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.Status"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Priority": {
            "type": "string",
            "enum": [
                "high",
                "normal",
                "low"
            ],
            "x-enum-varnames": [
                "PriorityHigh",
                "PriorityNormal",
                "PriorityLow"
            ]
        },
        "domain.Status": {
            "type": "string",
            "enum": [
                "pending",
                "queued",
                "processing",
                "delivered",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusQueued",
                "StatusProcessing",
                "StatusDelivered",
                "StatusFailed",
                "StatusCancelled"
            ]
        }
    }
}